+++ backend/Rhyolite/Backend/Sign.hs
@@ -23,19 +23,19 @@ import Control.Monad.Reader
 import Control.Monad.Trans.Control
 import Data.Aeson (FromJSON, ToJSON, encode)
 import qualified Data.ByteString.Lazy as LBS
-import Data.Proxy (Proxy (..))
+import Data.Proxy (Proxy(..))
 import Data.Text.Encoding (decodeUtf8, encodeUtf8)
 import Data.Typeable (Typeable, typeRep)
 import Database.Groundhog (DbPersist)
+import Database.PostgreSQL.Simple.Class
 import qualified Web.ClientSession as CS
 
 import Rhyolite.Backend.Schema.TH (deriveNewtypePersistBackend)
 import Rhyolite.Backend.DB.LargeObjects (PostgresLargeObject (withLargeObject))
-import Rhyolite.Backend.DB.PsqlSimple (Psql)
 import Rhyolite.Email (MonadEmail)
 import Rhyolite.Request.Common (decodeValue')
 import Rhyolite.Route (MonadRoute)
-import Rhyolite.Sign (MonadSign (..), Signed (..))
+import Rhyolite.Sign (MonadSign(..), Signed(..))
 
 signWithKey :: (Typeable b, ToJSON b, MonadIO m) => CS.Key -> b -> m (Signed a)
 signWithKey k (v :: b) =